wip: move templates to mystmd
