+++ packages/myst-cli/src/index.ts
@@ -4,6 +4,7 @@ export * from './config.js';
 export * from './init/index.js';
 export * from './frontmatter.js';
 export * from './plugins.js';
+export * from './templates.js';
 export * from './process/index.js';
 export * from './project/index.js';
 export * from './session/index.js';
